requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
blessings==1.7
brotlipy==0.7.0
click==8.1.7
cmake==3.26.3
conda==4.11.0
conda-build==3.21.7
ConfigArgParse==1.5.3
contourpy==1.0.6
cryptography==36.0.0
cycler==0.11.0
docker-pycreds==0.4.0
docstring_parser==0.16
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2022.11.0
functorch==0.1.1
gdown==4.7.1
gin-config==0.4.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==0.6.0
idna==3.3
imageio==2.27.0
imageio-ffmpeg==0.4.7
importlib-resources==5.12.0
jedi==0.18.1
Jinja2==3.0.3
jsonargparse==4.29.0
kiwisolver==1.4.4
kornia==0.7.2
kornia_rs==0.1.3
lazy_loader==0.2
lightning-utilities==0.8.0
lit==16.0.5
MarkupSafe==2.0.1
matplotlib==3.7.1
matplotlib-inline==0.1.3
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.3.0
multidict==6.0.2
networkx==3.1
ninja==1.11.1
numpy==1.21.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py3==7.352.0
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.7.0.72
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
piq==0.8.0
piqa==1.2.2
platformdirs==4.2.2
protobuf==3.20.3
psutil==5.8.0
pytorch-lightning==2.0.9.post0
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
requests==2.26.0
scikit-image==0.20.0
scipy==1.9.1
sentry-sdk==2.6.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.3.1
sympy==1.12
tensorboardX==2.6
termcolor==2.1.1
tifffile==2023.4.12
torch==1.11.0+cu113
torch-efficient-distloss==0.1.3
torch-scatter==2.1.2
torchaudio==0.11.0+cu113
torchmetrics==0.11.1
torchtext==0.12.0
torchvision==0.12.0+cu113
tqdm==4.62.3
triton==2.3.1
typeshed_client==2.5.1
typing_extensions==4.12.2
urllib3==1.26.7
wandb==0.17.2
wcwidth==0.2.5
yarl==1.8.1
zipp==3.15.0
